+++ src/instances.ts
@@ -152,7 +152,7 @@ function successfulTypeScriptInstance(
     }
 
     // if allowJs is set then we should accept js(x) files
-    const scriptRegex = configParseResult.options.allowJs && !loaderOptions.entryFileCannotBeJs
+    const scriptRegex = configParseResult.options.allowJs
         ? /\.tsx?$|\.jsx?$/i
         : /\.tsx?$/i;